* requirements.txt modified * Solution to duplicated id problem * Unsolved situation when the number of cases returned by the blast is not the square of the number of sequences

+++ calculate/blast.py
@@ -8,6 +8,8 @@ import os
 
 class BlastCalculation(QtCore.QThread):
     onfinished = QtCore.pyqtSignal(pd.DataFrame)
+    onincompletedata = QtCore.pyqtSignal(str)
+    onblastincomplete = QtCore.pyqtSignal(str)
 
     def __init__(self, matrix_type, fasta):
         QtCore.QThread.__init__(self)
@@ -17,45 +19,63 @@ class BlastCalculation(QtCore.QThread):
         self.cpucount = mp.cpu_count()
         self.type = {'bitscores': 'bitscore', 'raw-scores': 'score', 'identity': 'pident'}
 
-    def clean_duplicates(self, raw_data):
+    def modify_id(self, seqs):
+        for seqindex in range(len(seqs)):
+            seq = seqs[seqindex]
+            seq.id = '_'.join(seq.description.split())
+            seqs[seqindex] = seq
+        return seqs
+        #SeqIO.write(seqs, 'temp.fasta', 'fasta')
 
-        raw_data['check_string'] = raw_data.apply(lambda row: ''.join(sorted([row['seq1'], row['seq2']])), axis=1)
-        no_dups = raw_data.drop_duplicates('check_string')
-        no_dups.index = pd.RangeIndex(start=1, stop=len(no_dups.index) + 1)
-
-        return no_dups
+    # def clean_duplicates(self, raw_data):
+    #     raw_data['check_string'] = raw_data.apply(lambda row: ''.join(sorted([row['seq1'], row['seq2']])), axis=1)
+    #     no_dups = raw_data.drop_duplicates('check_string')
+    #     no_dups.index = pd.RangeIndex(start=1, stop=len(no_dups.index) + 1)
+    #     return no_dups
 
     def run(self):
-        num_cases = len(list(SeqIO.parse(self.fasta, 'fasta')))
+        seqs = list(SeqIO.parse(self.fasta, 'fasta'))
+        num_cases = len(list(seqs))
+        unique_cases = len(set([seq.id for seq in seqs]))
 
-        # Make Database
-        clinedb = NcbimakeblastdbCommandline(cmd='makeblastdb', dbtype='prot', input_file=self.fasta,
-                                             input_type='fasta', out=self.fasta)
-        clinedb()
+        if num_cases > unique_cases:
+            seqs = self.modify_id(seqs)
+            self.fasta = 'temp.fasta'
+            actual_cases = len(set([seq.id for seq in seqs]))
 
-        # Calculation
-        clinec = NcbiblastpCommandline(cmd='blastp', query=self.fasta, db=self.fasta, evalue=10,
-                                       outfmt='6 qseqid sseqid pident evalue bitscore score', max_target_seqs=num_cases,
-                                       max_hsps=1, num_threads=self.cpucount, out='all_vs_all.tsv')
-        clinec()
+            if num_cases == actual_cases:
+                # Make Database
+                clinedb = NcbimakeblastdbCommandline(cmd='makeblastdb', dbtype='prot', input_file=self.fasta,
+                                                     input_type='fasta', out=self.fasta)
+                clinedb()
 
-        # Data Processing
-        data = pd.read_csv('all_vs_all.tsv', delimiter='\t', names=['seq1', 'seq2', 'pident', 'evalue', 'bitscore',
-                                                                    'score'])
-        no_dups = self.clean_duplicates(data)
-        pivoted = no_dups.pivot(index='seq1', columns='seq2', values=self.type[self.matrix_type])
+                # Calculation
+                clinec = NcbiblastpCommandline(cmd='blastp', query=self.fasta, db=self.fasta, evalue=0.1,
+                                               outfmt='6 qseqid sseqid pident evalue bitscore score',
+                                               max_target_seqs=num_cases, max_hsps=1, num_threads=self.cpucount,
+                                               out='all_vs_all.tsv')
+                clinec()
 
-        for column in pivoted.columns:
-            pos = pivoted.index.get_loc(column) + 1
-            for index in pivoted.index[pos:]:
-                if column != index:
-                    pivoted.loc[index, column] = pivoted.loc[column, index]
+                # Data Processing
+                data = pd.read_csv('all_vs_all.tsv', delimiter='\t', names=['seq1', 'seq2', 'pident', 'evalue', 'bitscore',
+                                                                            'score'])
+                pivoted = data.pivot(index='seq1', columns='seq2', values=self.type[self.matrix_type])
 
-        os.remove('all_vs_all.tsv')
-        os.remove('{}.phr'.format(self.fasta))
-        os.remove('{}.pin'.format(self.fasta))
-        os.remove('{}.psq'.format(self.fasta))
+                for column in pivoted.columns:
+                    pos = pivoted.index.get_loc(column) + 1
+                    for index in pivoted.index[pos:]:
+                        if column != index:
+                            pivoted.loc[index, column] = pivoted.loc[column, index]
 
-        pivoted_round = pivoted.round(2)
+                os.remove('all_vs_all.tsv')
+                os.remove('{}.phr'.format(self.fasta))
+                os.remove('{}.pin'.format(self.fasta))
+                os.remove('{}.psq'.format(self.fasta))
 
-        self.onfinished.emit(pivoted_round)
+                pivoted_round = pivoted.round(2)
+
+                self.onfinished.emit(pivoted_round)
+            elif num_cases > actual_cases:
+                self.onincompletedata.emit('Duplicate sequences in "fasta"')
+            else:
+                self.onblastincomplete.emit("El blast no completó el total de los casos")
